fix warning: format ‘%d’ expects type ‘int’, but argument 5 has type ‘integer’
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@2181 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/tex/textoken.c
@@ -503,7 +503,7 @@ void check_outer_validity(void)
             }
             snprintf(errmsg, 255,
                      "Incomplete %s; all text was ignored after line %d",
-                     cmd_chr_to_string(if_test_cmd, cur_if), skip_line);
+                     cmd_chr_to_string(if_test_cmd, cur_if), (int)skip_line);
             /* @.Incomplete \\if...@> */
             cur_tok = cs_token_flag + frozen_fi;
             /* back up one inserted token and call |error| */
